backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
sqlalchemy==2.0.23
alembic==1.13.1
psycopg2-binary==2.9.9
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
google-generativeai==0.3.2
pandas==2.1.4
numpy==1.26.4
scikit-learn==1.3.2
pydub==0.25.1
python-dotenv==1.0.0
SpeechRecognition==3.10.0
pytesseract==0.3.10
Pillow==10.2.0
